requirements.txt
attrs==24.2.0
blinker==1.8.2
click==8.1.7
flasgger==0.9.7.1
Flask==3.0.3
Flask-Cors==5.0.0
flask-marshmallow==1.2.1
Flask-SQLAlchemy==3.1.1
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==2.1.5
marshmallow==3.23.1
marshmallow-sqlalchemy==1.1.0
mistune==3.0.2
packaging==24.1
PyMySQL==1.1.1
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.0.4
zipp==3.20.2
